sync typdefs with newlib + cleanup

+++ include/fs/device.h
@@ -21,7 +21,7 @@ struct device
 	enum device_minor minor;
 };
 
-typedef uint32_t dev_t;
+typedef int16_t dev_t;
 typedef uint32_t mode_t;
 
 #endif
