+++ include/kernel/types.h
@@ -11,8 +11,8 @@
 typedef uintptr_t p_addr_t; // physical address
 typedef uintptr_t v_addr_t; // virtual address
 
-typedef long long off_t;
+typedef long off_t;
 
-typedef int16_t pid_t;
+typedef int pid_t;
 
 #endif
